+++ src/HLS/binding/module/cdfc_module_binding.hpp
@@ -12,7 +12,7 @@
  *                       Politecnico di Milano - DEIB
  *                        System Architectures Group
  *             ***********************************************
- *              Copyright (C) 2004-2022 Politecnico di Milano
+ *              Copyright (C) 2004-2023 Politecnico di Milano
  *
  *   This file is part of the PandA framework.
  *
@@ -401,7 +401,7 @@ class cdfc_module_binding : public fu_binding_creator
 #ifdef HC_APPROACH
                           spec_hierarchical_clustering& hc,
 #endif
-                          connection_relation& con_rel, double controller_delay, unsigned int prec);
+                          connection_relation& con_rel, double controller_delay, unsigned long long prec);
 
    void update_slack_starting_time(const OpGraphConstRef fdfg, OpVertexSet& sorted_vertices,
                                    CustomUnorderedMap<vertex, double>& slack_time,
